350. Intersection of Two Arrays II.
